requirements.txt
Flask==3.0.0
python-docx==1.1.0
google-generativeai==0.3.2
gunicorn==21.2.0
werkzeug==3.0.1